requirements.txt
paddleocr
opencv-python
numpy
matplotlib
Pillow
torch
transformers
openai 